Refactor auth flow and update routing structure
Moved AuthFlow component and related logic from /auth to /authflow directory. Updated /auth/page.tsx to use SigninForm instead of AuthFlow, and created a new /authflow/page.tsx to handle the AuthFlow logic. This improves separation of concerns and clarifies the authentication flow routing.
